backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
pillow==12.0.0
psycopg==3.2.12
psycopg2-binary==2.9.11
python-dotenv==1.2.1
sqlparse==0.5.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-extensions==3.2.3
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
django-storages==1.14.2
boto3==1.34.0
channels==4.0.0
channels-redis==4.2.0
daphne==4.1.0

# Testing
pytest==8.3.4
pytest-django==4.9.0
pytest-asyncio==0.24.0
factory_boy==3.3.0
coverage==7.4.0
